created nested routes in App.js for shared layouts
